+++ src/Sulu/Bundle/ContactBundle/Controller/ContactController.php
@@ -446,7 +446,7 @@ class ContactController extends RestController implements ClassResourceInterface
     protected function getContactsByUserSystem()
     {
         $repo = $this->get('sulu_security.user_repository');
-        $users = $repo->getUserInSystem();
+        $users = $repo->findUserBySystem($this->getParameter('sulu_security.system'));
         $contacts = [];
 
         foreach ($users as $user) {